Extract reusable stuff to Pro package
