+++ packages/x-data-grid-premium/src/hooks/features/rows/useGridRowsOverridableMethods.ts
@@ -5,12 +5,14 @@ import {
   gridExpandedSortedRowIdsSelector,
   gridRowNodeSelector,
 } from '@mui/x-data-grid-pro';
-import { gridExpandedSortedRowIndexLookupSelector } from '@mui/x-data-grid-pro/internals';
+import {
+  gridExpandedSortedRowIndexLookupSelector,
+  type ReorderExecutionContext,
+} from '@mui/x-data-grid-pro/internals';
 import type { RefObject } from '@mui/x-internals/types';
-import { rowGroupingReorderExecutor } from '../rowReorder/reorderExecutor';
+import { rowGroupingReorderExecutor } from '../rowReorder/rowGroupingReorderExecutor';
 import type { GridPrivateApiPremium } from '../../../models/gridApiPremium';
 import type { DataGridPremiumProcessedProps } from '../../../models/dataGridPremiumProps';
-import type { ReorderExecutionContext } from '../rowReorder/types';
 
 export const useGridRowsOverridableMethods = (
   apiRef: RefObject<GridPrivateApiPremium>,
